Added changes to support creation of new hp profile and added dto's for personal, workprofile, registration details.

+++ src/main/java/in/gov/abdm/nmr/repository/IUniversityRepository.java
@@ -1,12 +1,13 @@
 package in.gov.abdm.nmr.repository;
 
+import java.math.BigInteger;
 import java.util.List;
 
 import in.gov.abdm.nmr.entity.University;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 
-public interface IUniversityRepository extends JpaRepository<University, Long> {
+public interface IUniversityRepository extends JpaRepository<University, BigInteger> {
 
     @Query(value = "SELECT name, id, location FROM universities", nativeQuery = true)
     List<University> getUniversity();
